# src/api/openapi.yaml
openapi: 3.0.0
info:
  description: LeBe Solarenergie — Product Administration API
  version: 1.0.0
  title: LeBe Solarenergie Administration API
components:
  schemas:
...
    Product:
      type: object
      required: [id, category, Header, manufacturer_id, Status]
      properties:
        id:
          type: string
          example: MOD-001
        category:
          type: string
          enum: [Solarmodule, Wechselrichter, Heimspeicher, Ladestationen, Heizsysteme]
        Header:
          type: string
        Beschreibung:
          type: string
        Hersteller:
          type: string
        manufacturer_id:
          type: string
        Garantie:
          type: string
        Power:
          type: number
          nullable: true
        Unit:
          type: string
        Spezifikation:
          type: string
          nullable: true
          description: URL of the uploaded specification PDF
        hasSpec:
          type: boolean
        Logo:
          type: string
          nullable: true
        Status:
          type: string
          enum: [Active, Draft, Hidden]
        panelHeightMeters:
          type: number
          nullable: true
        panelWidthMeters:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  parameters:
    id:
      in: path
      required: true
      name: id
      description: Resource id
      schema:
        type: string
paths:
...
  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: manufacturer_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      summary: Create a product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Created
        "422":
          description: Validation error
  /products/next-id:
    get:
      summary: Generate the next auto-numbered product id for a category
      parameters:
        - in: query
          name: category
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
  /products/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get a product
      responses:
        "200":
          description: Success
        "404":
          description: Not found
    put:
      summary: Update a product (renaming the id recreates the document)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Success
        "422":
          description: Validation error
    delete:
      summary: Delete a product
      responses:
        "204":
          description: Deleted
...